bot/requirements.txt
backtrader==1.9.76.123
numpy==1.21.0
pandas==1.3.2
python_binance==1.0.12
binance==0.3
dataclasses==0.6
PyYAML==6.0
tabulate==0.8.9
python_-telegram-bot==13.8.1
termcolor==1.1.0
tqdm==4.62.3

pymongo~=3.12.1